requirements.txt
flask==2.3.2
flask-mysqldb==1.0.1
python-dotenv==1.0.0
mysql-connector-python==9.2.0
pyinstaller==6.12.0
